requirements.txt
ultralytics>=8.3.0
streamlit>=1.35.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
